mapa en el home, mostrar nmás productos en home y productos
