+++ Client/src/components/Home/Home.component.jsx
@@ -3,8 +3,9 @@ import React, { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 //COMPONENTS
 import HomeProducts from "../HomeProducts/HomeProducts.component";
-import CarouselBottomComponent from "../../components/CarouselBottom/CarouselBottom.component";
-import CarouselTopComponent from "../../components/CarouselTop/CarouselTop.component";
+import CarouselBottom from "../../components/CarouselBottom/CarouselBottom.component";
+import CarouselTop from "../../components/CarouselTop/CarouselTop.component";
+import GoogleMap from "../../components/GoogleMap/GoogleMap.component";
 //REDUX
 import { fetchAllProducts } from "../../services/ProductServices";
 
@@ -19,9 +20,10 @@ const HomeComponent = () => {
   return (
     <>
       <div>
-        <CarouselTopComponent />
+        <CarouselTop />
         <HomeProducts allProducts={allProducts} />
-        <CarouselBottomComponent />
+        <GoogleMap/>
+        <CarouselBottom />
       </div>
     </>
   );
